molayo_backend_FastAPI/pyproject.toml
[tool.poetry]
name = "molayo-backend-fastapi"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.10"
fastapi = {extras = ["all"], version = "^0.105.0"}
uvicorn = {extras = ["standard"], version = "^0.25.0"}
sqlmodel = "^0.0.14"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
password-validator = "^1.0"
bcrypt = "4.0.1"
aiofiles = "^23.2.1"
tensorflow = "2.10.1"
tensorflow-io-gcs-filesystem = "0.31.0"
pillow = "^10.2.0"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
